Generate and send weekly outcomes report to each client every Monday morning
Every Monday 11 A.M messages.api.ts runs weeklyReport(). WeeklyReport then finds all patients that have outcomes from last week. Then, for each patient it calculates the data required by the message template and creates the templates. Then it create the message and saves it so the Twilio scheduler takes the message and sends it.